+++ src/image/mod.rs
@@ -1,0 +1,1 @@
+pub mod png_image;
