+++ src/utils.rs
@@ -1,0 +1,5 @@
+
+#[inline]
+pub fn u8_to_u4(x: u8) -> u8 {
+    x >> 4
+}
